requirements.txt
Flask==2.3.3
python-telegram-bot==20.6
SQLAlchemy==2.0.23
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
requests==2.31.0
gunicorn==21.2.0 